fix: address CodeRabbit and Gemini review
ci: do not persist the checkout credential in the oracle job. That job
builds FFmpeg from source, so external build scripts run in the
workspace and a token sitting in the git config is needless exposure.
Matches what the fuzz job already does.

ci: drain the oracle capability checks instead of closing the pipe
early. Under pipefail, head -1 and grep -q can close the pipe and kill
ffmpeg with SIGPIPE, failing the step even though the oracle is fine.
This is latent rather than live: -decoders emits 25892 bytes, which fits
the 64K pipe buffer, so ffmpeg finishes writing and it does not fire
today. The margin is only 2.5x and the failure would surface as a
baffling intermittent red, so drain anyway.

hooks: fail closed when git log cannot inspect the range. 2>/dev/null
plus || true turned an unreadable range into an empty result, and an
empty result reads as "no bad identities", so a broken query waved the
push through. That is precisely the silent-pass failure this guard
exists to prevent, and the same shape as the skipped-gate bug this
branch fixes.

hooks: exclude only commits already on the destination remote. Plain
--remotes ignores commits reachable from ANY configured remote, so a
commit living only on a different remote could slip past on a new
branch.

tools/cdec: cast n to size_t and flen to int explicitly, plus the same
conversion one line up in fread, which -Wconversion also flags.

Verified: the hook refuses a work-email commit, refuses an uninspectable
range, and passes a clean branch; cdec still reproduces all 10 golden
dumps byte-identically.

+++ tools/cdec/cdec.c
@@ -143,9 +143,9 @@ static uint8_t *read_all(const char *path, size_t *size)
         fprintf(stderr, "%s: cannot allocate %ld bytes\n", path, n);
         exit(1);
     }
-    if (fread(buf, 1, n, f) != (size_t)n) { perror("read"); exit(1); }
+    if (fread(buf, 1, (size_t)n, f) != (size_t)n) { perror("read"); exit(1); }
     fclose(f);
-    *size = n;
+    *size = (size_t)n;
     return buf;
 }
 
@@ -234,7 +234,7 @@ int main(int argc, char **argv)
         }
         fprintf(g_out, "FRAME %d\n", fidx);
         av_packet_unref(pkt);
-        if (av_new_packet(pkt, flen) < 0) {
+        if (av_new_packet(pkt, (int)flen) < 0) {
             fprintf(stderr, "cdec: av_new_packet(%zu) failed at frame %d\n",
                     flen, fidx);
             fclose(g_out);
